backend/requirements.txt
streamlit>=1.28.0
musicai_sdk>=1.0.3
pydub>=0.25.1
soundfile>=0.12.1
python-dotenv>=1.0.0